Fix: book mock bug not using book factory
When creating a mock book obj for testing I created an object with no
getter/setter functions allowing for direct mutation of the objects key
value pairs. I corrected this by requiring the book module and creating
a mocked book obj using the book module which is more representative of
the application logic. Additionally I reformatted the `addBook` and
`deleteBook` functions to utilize new mock book.